GifPicker has functioning search
The Express API now supports searching for GIFs via the Tenor API.

Searched GIFs are displayed in the GIF picker. The GIFs are displayed horizontally, rather than vertically with 2
columns, like the Masonry MUI component. This is unintentional and needs
to be fixed.

The search functionality will need to be optimized to prevent too many
API calls to our API as well as the Tenor API.